Added Utility::EscapedString class.
This class allows lazy transformation of the given source string by substituting
set of "forbidden" characters with escape symbol and other substituted character.
The main idea is to avoid unwanted memory allocation: when it is possible no
memory allocation performed. To achieve this lazy evaluation is used: string
transformation is perfomed only if it is needed, and memory allocation
performed only in cases when it is unavoidable.